src/test/java/guru/springframework/controllers/PropertyInjectedControllerTest.java
package guru.springframework.controllers;

import guru.springframework.services.impl.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.assertEquals;

public class PropertyInjectedControllerTest {

    PropertyInjectedController propertyInjectedController;

    @Before
    public void setUp() throws Exception{
        this.propertyInjectedController=new PropertyInjectedController();
        this.propertyInjectedController.propertyGreetingServiceImpl=new GreetingServiceImpl();
    }

    @Test
    public void testGreeting() throws Exception{
        assertEquals(GreetingServiceImpl.HELLOWORLD,propertyInjectedController.hello());
    }
}
